:truck: chore: move login to pages directory
