+++ src/app/modules/shared/shared-module.ts
@@ -8,7 +8,7 @@ import { NavbarComponent } from './components/navbar/navbar';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { ButtonComponent } from './components/button/button';
 import { InputComponent } from './components/input/input';
-import { Login } from './components/login/login';
+import { Login } from './pages/login/login';
 
 
 @NgModule({
@@ -31,7 +31,8 @@ import { Login } from './components/login/login';
     Pagination,
     Radios,
     ButtonComponent,
-    NavbarComponent
+    NavbarComponent,
+    Login
   ]
 })
 export class SharedModule { }
